requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-phonenumber-field==6.1.0
phonenumbers==8.12.47
Pillow==9.1.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
